# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TotalEasy API
  contact:
    email: marcosvisentini@gmail.com
  license:
    name: GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.html
  version: v1
servers:
  - url: http://localhost/api
    description: Desenvolvimento
...
paths:
  /usuarios:
    get:
      tags:
        - usuarios
      summary: 'Encontra todos os usuários.'
      description: 'Encontra todos os usuários.'
      operationId: findUsuarios
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
    post:
      tags:
        - usuarios
      summary: 'Cria um usuário.'
      description: 'Cria um usuário.'
      operationId: createUsuario
      requestBody:
        description: 'Os dados necessários para criar um usuário.'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioCreationDTO'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
  /processos-eleitorais:
    get:
      tags:
        - processos-eleitorais
      summary: 'Encontra todos os processos eleitorais.'
      description: 'Encontra todos os processos eleitorais.'
      operationId: findProcessosEleitorais
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProcessoEleitoralRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
...
  /papeis:
    get:
      tags:
        - papeis
      summary: 'Encontra todos os papéis.'
      description: 'Encontra todos os papéis.'
      operationId: findPapeis
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PapelRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
  /papeis/{nome}:
    get:
      tags:
        - papeis
      summary: 'Encontra um papel.'
      description: 'Encontra um papel.'
      operationId: findPapel
      parameters:
        - name: nome
          in: path
          description: 'O nome do papel.'
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 15
            example: DBA
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PapelRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
  /papeis/{nome}/usuarios:
    get:
      tags:
        - papeis
      summary: 'Encontra todos os usuários que possuem determinado papel.'
      description: 'Encontra todos os usuários que possuem determinado papel.'
      operationId: findUsuariosPapel
      parameters:
        - name: nome
          in: path
          description: 'O nome do papel.'
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 15
            example: DBA
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
  /papeis/{nome}/permissoes:
    get:
      tags:
        - papeis
      summary: 'Encontra todas as permissões que um papel possui.'
      description: 'Encontra todas as permissões que um papel possui.'
      operationId: findPermissoesPapel
      parameters:
        - name: nome
          in: path
          description: 'O nome do papel.'
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 15
            example: DBA
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermissaoRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
  /permissoes:
    get:
      tags:
        - permissoes
      summary: 'Encontra todas as permissões.'
      description: 'Encontra todas as permissões.'
      operationId: findPermissoes
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermissaoRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
...
  /origens-configuracao-processo-eleitoral:
    get:
      tags:
        - origens-configuracao-processo-eleitoral
      summary: 'Encontra todas as origens de configuração de processo eleitoral.'
      description: 'Encontra todas as origens de configuração de processo eleitoral.'
      operationId: findOrigensConfiguracaoProcessoEleitoral
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrigemConfiguracaoProcessoEleitoralRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
  /origens-configuracao-processo-eleitoral/{nomeAbreviado}:
    get:
      tags:
        - origens-configuracao-processo-eleitoral
      summary: 'Encontra uma origem de configuração de processo eleitoral.'
      description: 'Encontra uma origem de configuração de processo eleitoral.'
      operationId: findOrigemConfiguracaoProcessoEleitoral
      parameters:
        - name: nomeAbreviado
          in: path
          description: 'O nome abreviado da origem de configuração de processo eleitoral.'
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 3
            example: LEG
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrigemConfiguracaoProcessoEleitoralRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
  /origens-configuracao-processo-eleitoral/{nomeAbreviado}/processos-eleitorais:
    get:
      tags:
        - origens-configuracao-processo-eleitoral
      summary: 'Encontra todos os processos eleitorais com determinada origem de configuração.'
      description: 'Encontra todos os processos eleitorais com determinada origem de configuração.'
      operationId: findProcessosEleitoraisOrigemConfiguracaoProcessoEleitoral
      parameters:
        - name: nomeAbreviado
          in: path
          description: 'O nome abreviado da origem de configuração de processo eleitoral.'
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 3
            example: LEG
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProcessoEleitoralRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
...
  /candidaturas:
    get:
      tags:
        - candidaturas
      summary: 'Encontra uma ou mais candidaturas.'
      description: 'Se o parâmetro `id` for informado, é retornada uma candidatura identificado por ele. Se não, são retornadas todas as candidaturas.'
      operationId: findCandidaturas
      parameters:
        - in: query
          name: id
          description: 'Os dados que identificam a candidatura.'
          schema:
            $ref: '#/components/schemas/CandidaturaIdDTO'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CandidaturaRetrievalDTOArray'
                  - $ref: '#/components/schemas/CandidaturaRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
...
  /partidos:
    get:
      tags:
        - partidos
      summary: 'Encontra todos os partidos.'
      description: 'Encontra todos os partidos.'
      operationId: findPartidos
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PartidoRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
  /partidos/{numeroTSE}:
    get:
      tags:
        - partidos
      summary: 'Encontra um partido.'
      description: 'Encontra um partido.'
      operationId: findPartido
      parameters:
        - name: numeroTSE
          in: path
          description: 'O número do partido.'
          required: true
          schema:
            type: integer
            format: int64
            example: 11
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartidoRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
  /partidos/{numeroTSE}/candidaturas:
    get:
      tags:
        - partidos
      summary: 'Encontra todas as candidaturas de um partido.'
      description: 'Encontra todas as candidaturas de um partido.'
      operationId: findCandidaturasPartido
      parameters:
        - name: numeroTSE
          in: path
          description: 'O número do partido.'
          required: true
          schema:
            type: integer
            format: int64
            example: 11
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CandidaturaRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
  /partidos/{numeroTSE}/apuracoes-votos-boletim-urna:
    get:
      tags:
        - partidos
      summary: 'Encontra todas as apurações de votos por boletim de urna de um partido.'
      description: 'Encontra todas as apurações de votos por boletim de urna de um partido.'
      operationId: findApuracoesVotosPartidoBoletimUrna
      parameters:
        - name: numeroTSE
          in: path
          description: 'O número do partido.'
          required: true
          schema:
            type: integer
            format: int64
            example: 11
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApuracaoVotosPartidoBoletimUrnaRetrievalDTO'
        '4XX':
          description: 'Client error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
      security:
        - bearerAuth: []
...
components:
  schemas:
    UsuarioRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        username:
          type: string
          example: wisentini
        nome:
          type: string
          example: Marcos
        sobrenome:
          type: string
          example: Visentini
        papeis:
          type: array
          items:
            type: string
          example: [ADMIN, USER]
        criadoEm:
          type: string
          format: date-time
          example: 03/03/2023 09:29:16
        atualizadoEm:
          type: string
          format: date-time
          example: 12/05/2023 14:05:23
...
    ZonaRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 34
        numeroTSE:
          type: integer
          format: int64
          example: 18
        uf:
          $ref: '#/components/schemas/UFRetrievalDTO'
    SecaoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 61
        numeroTSE:
          type: integer
          format: int64
          example: 112
        zona:
          $ref: '#/components/schemas/ZonaRetrievalDTO'
    OrigemConfiguracaoProcessoEleitoralRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          example: 'Eleição legal oficial'
        nomeAbreviado:
          type: string
          example: LEG
    ProcessoEleitoralRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        codigoTSE:
          type: integer
          format: int64
          example: 395
        nome:
          type: string
          example: 'Eleições Gerais 2022'
        origemConfiguracao:
          $ref: '#/components/schemas/OrigemConfiguracaoProcessoEleitoralRetrievalDTO'
    PleitoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        codigoTSE:
          type: integer
          format: int64
          example: 406
        nome:
          type: string
          example: '1º turno'
        turno:
          type: integer
          format: int64
          example: 1
        data:
          type: string
          format: date
          example: 02/10/2022
        processoEleitoral:
          $ref: '#/components/schemas/ProcessoEleitoralRetrievalDTO'
    SecaoPleitoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 61
        secao:
          $ref: '#/components/schemas/SecaoRetrievalDTO'
        pleito:
          $ref: '#/components/schemas/PleitoRetrievalDTO'
    FaseRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        codigoTSE:
          type: integer
          format: int64
          example: 2
        nome:
          type: string
          example: Oficial
    GeneroRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        codigoTSE:
          type: integer
          format: int64
          example: 2
        nome:
          type: string
          example: MASCULINO
    OcupacaoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        codigoTSE:
          type: integer
          format: int64
          example: 277
        nome:
          type: string
          example: 'DEPUTADO'
    GrauInstrucaoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        codigoTSE:
          type: integer
          format: int64
          example: 8
        nome:
          type: string
          example: 'SUPERIOR COMPLETO'
    CorRacaRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        codigoTSE:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          example: BRANCA
    EstadoCivilRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        codigoTSE:
          type: integer
          format: int64
          example: 3
        nome:
          type: string
          example: 'CASADO(A)'
    OrigemBoletimUrnaRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          example: 'Software de Votação'
        nomeAbreviado:
          type: string
          example: Vota
    UrnaEletronicaRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 61
        numeroSerie:
          type: integer
          format: int64
          example: 1303045
        codigoIdentificacaoCarga:
          type: string
          example: '904585283618888740610421'
        versaoSoftware:
          type: string
          example: '8.26.0.0'
        dataAbertura:
          type: string
          format: date
          example: 02/10/2022
        horarioAbertura:
          type: string
          format: time
          example: 08:00:01
        dataFechamento:
          type: string
          format: date
          example: 02/10/2022
        horarioFechamento:
          type: string
          format: time
          example: 17:04:48
    BoletimUrnaRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 61
        secaoPleito:
          $ref: '#/components/schemas/SecaoPleitoRetrievalDTO'
        fase:
          $ref: '#/components/schemas/FaseRetrievalDTO'
        origem:
          $ref: '#/components/schemas/OrigemBoletimUrnaRetrievalDTO'
        urnaEletronica:
          $ref: '#/components/schemas/UrnaEletronicaRetrievalDTO'
        assinatura:
          type: string
          example: A68CF51ECA1A79BB80D5514012AB8D2E7424291023FB42E0CFEB4849D7337720434E8D30C798230B5695D1ADF19B2142AAC1687A1C6D81BC37A406041E238D00
        quantidadeTotalQRCodes:
          type: integer
          format: int64
          example: 3
        dataEmissao:
          type: string
          format: date
          example: 02/10/2022
        horarioEmissao:
          type: string
          format: time
          example: 17:05:53
...
    PapelRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 4
        nome:
          type: string
          example: ADMIN
        descricao:
          type: string
          example: 'Administrador'
    PermissaoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        nome:
          type: string
          example: ESCRITA
        descricao:
          type: string
          example: 'Acesso para escrita'
...
    TipoCargoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        codigoTSE:
          type: integer
          format: int64
          example: 1
        nome:
          type: string
          example: 'Proporcional'
    CargoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        codigoTSE:
          type: integer
          format: int64
          example: 6
        nome:
          type: string
          example: 'Deputado Federal'
        nomeAbreviado:
          type: string
          example: 'Dep. Fed.'
        tipo:
          $ref: '#/components/schemas/TipoCargoRetrievalDTO'
    EleicaoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        codigoTSE:
          type: integer
          format: int64
          example: 546
        nome:
          type: string
          example: 'Eleições Gerais Estaduais 2022'
        ano:
          type: integer
          format: int64
          example: 2022
        quantidadeVotosCargosMajoritarios:
          type: integer
          format: int64
          example: 197
        quantidadeVotosCargosProporcionais:
          type: integer
          format: int64
          example: 64
        pleito:
          $ref: '#/components/schemas/PleitoRetrievalDTO'
    CargoEleicaoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        cargo:
          $ref: '#/components/schemas/CargoRetrievalDTO'
        eleicao:
          $ref: '#/components/schemas/EleicaoRetrievalDTO'
    CandidatoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 16761
        numeroTSE:
          type: integer
          format: int64
          example: 1166
        codigoTSE:
          type: string
          example: '210001646689'
        nomeCompleto:
          type: string
          example: 'JOSÉ ALFONSO EBERT HAMM'
        nomeUrna:
          type: string
          example: 'AFONSO HAMM'
        nomeSocial:
          type: string
          example: 'AFONSO'
        numeroCPF:
          type: string
          example: '37040642034'
        numeroTituloEleitoral:
          type: string
          example: '011694010477'
        dataNascimento:
          type: string
          format: date
          example: '25/04/1962'
        genero:
          $ref: '#/components/schemas/GeneroRetrievalDTO'
        corRaca:
          $ref: '#/components/schemas/CorRacaRetrievalDTO'
        grauInstrucao:
          $ref: '#/components/schemas/GrauInstrucaoRetrievalDTO'
        estadoCivil:
          $ref: '#/components/schemas/EstadoCivilRetrievalDTO'
        ocupacao:
          $ref: '#/components/schemas/OcupacaoRetrievalDTO'
    PartidoRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        numeroTSE:
          type: integer
          format: int64
          example: 11
        nome:
          type: string
          example: PROGRESSISTAS
        sigla:
          type: string
          example: PP
    CandidaturaRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        candidato:
          $ref: '#/components/schemas/CandidatoRetrievalDTO'
        cargoEleicao:
          $ref: '#/components/schemas/CargoEleicaoRetrievalDTO'
        partido:
          $ref: '#/components/schemas/PartidoRetrievalDTO'
...
    ApuracaoVotosPartidoBoletimUrnaRetrievalDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 3192
        partido:
          $ref: '#/components/schemas/PartidoRetrievalDTO'
        boletimUrna:
          $ref: '#/components/schemas/BoletimUrnaRetrievalDTO'
        cargoEleicao:
          $ref: '#/components/schemas/CargoEleicaoRetrievalDTO'
        quantidadeVotosDeLegenda:
          type: integer
          format: int64
          example: 4
        totalVotosApurados:
          type: integer
          format: int64
          example: 91
...
    UsuarioCreationDTO:
      type: object
      required:
        - username
        - senha
        - nomesPapeis
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 31
          example: teste
        senha:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          example: teste123
        nome:
          type: string
          minLength: 1
          maxLength: 127
          example: 'João'
        sobrenome:
          type: string
          minLength: 1
          maxLength: 127
          example: 'da Silva'
        nomesPapeis:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 15
          example: [USER]
...
    CandidaturaIdDTO:
      type: object
      required:
        - numeroTSECandidato
        - codigoTSECargo
        - codigoTSEEleicao
      properties:
        numeroTSECandidato:
          type: integer
          format: int64
          example: 1166
        codigoTSECargo:
          type: integer
          format: int64
          example: 6
        codigoTSEEleicao:
          type: integer
          format: int64
          example: 546
...
    ApiResponse:
      type: object
      required:
        - status
        - content
        - timestamp
      properties:
        status:
          type: string
          example: OK
        content:
          type: string
          example: 'Conteúdo da resposta'
        timestamp:
          type: string
          format: date-time
          example: '03/03/2023 16:25:00'
    ApiError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: 'Descrição detalhada de um erro'
    ApiErrorResponse:
      type: object
      required:
        - status
        - errors
        - timestamp
      properties:
        status:
          type: string
          example: BAD_REQUEST
        description:
          type: string
          example: 'Descrição genérica de um ou mais erros'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ApiError'
        timestamp:
          type: string
          format: date-time
          example: '05/03/2023 09:43:11'
...
    CandidaturaRetrievalDTOArray:
        type: array
        items:
          $ref: '#/components/schemas/CandidaturaRetrievalDTO'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT